Continue generator decomposition for CLI command handlers
